rename: 도메인 명을 user에서 member로 변경
